fin SO3

+++ S03TipsCalculator/app/src/main/java/com/example/tipscalculator/domain/models/TipsCalculation.kt
@@ -5,7 +5,7 @@ import com.example.tipscalculator.core.Constants
 
 data class TipsCalculation(private val subtotal: Double, private val tipsPercent: Int) {
 
-    val tipsAmount = subtotal * (tipsPercent / Constants.PERCENT)
+    val tipsAmount = subtotal * (tipsPercent.toFloat() / Constants.PERCENT)
     val TPS = subtotal * Constants.TPS
     val TVQ = subtotal * Constants.TVQ
     val total = subtotal + TPS + TVQ + tipsAmount
